+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 + Fluid interaction has potion effects now
 + You can no longer crush stone to cobblestone in the StoneWork Factory
 + Fixed Dye Mixer not accepting white lenses
++ Lenses can now be recolored
 
 # 1.10.1
 + If the black hole controller is lock the black hole units won't loose the filter